requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
trafilatura>=1.12.2
tldextract>=5.1.2
networkx>=3.3
pandas>=2.2.2
numpy>=2.0.2
scikit-learn>=1.5.2
sentence-transformers>=3.0.1
umap-learn>=0.5.6
pyvis>=0.3.2
tqdm>=4.66.5
readability-lxml>=0.8.1
streamlit>=1.39.0
